pyproject.toml
[tool.poetry]
name = "vook_db_v3"
version = "0.1.0"
description = ""
authors = ["atsushimemet <noap3b69n@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^3.5.2"
jupyterlab-code-formatter = "^1.5.3"
flake8 = "^6.0.0"
black = "^22.12.0"
isort = "^5.11.4"
mypy = "^0.991"
logzero = "^1.7.0"
google-cloud-bigquery = "^3.4.2"
pandas = "^1.5.3"
ipykernel = "^6.20.2"
pyarrow = "^10.0.1"
pandas-gbq = "^0.19.0"
types-requests = "^2.28.11"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
